media-analyzer/backend/requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
channels==4.0.0
channels-redis==4.2.0
celery==5.3.4
redis==5.0.1
ffmpeg-python==0.2.0
Pillow==10.0.1
django-cors-headers==4.3.1
torch==2.1.0
torchvision==0.16.0
transformers==4.36.0
opencv-python==4.8.1.78
numpy==1.24.3